Split file reference store interface into queries and updates
